fix: show Dolphin warning once per session
